Rename translator.py to query_ast_visitor.py
